# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for the Delivery Backend",
        "title": "Delivery Backend API",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "version": "1.0"
    },
    "host": "localhost:1317",
    "basePath": "/",
    "paths": {
        "/api/categories": {
            "get": {
                "description": "Retrieve a list of all dish categories",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories"
                ],
                "summary": "Get all dish categories",
                "responses": {
                    "200": {
                        "description": "List of categories",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve categories",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            },
            "post": {
                "description": "Add a new dish category by providing its name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories (Admin)"
                ],
                "summary": "Add a new dish category",
                "parameters": [
                    {
                        "description": "Category data",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.AddDishCategoryPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "ID of the created category",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "403": {
                        "description": "Access forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "500": {
                        "description": "Failed to create category",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            }
        },
        "/api/dishes": {
            "get": {
                "description": "Retrieve a list of all available dishes",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dishes"
                ],
                "summary": "Get all dishes",
                "responses": {
                    "200": {
                        "description": "List of dishes",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Dish"
                            }
                        }
                    },
                    "404": {
                        "description": "No dishes found",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve dishes",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            }
        },
        "/api/dishes/admin/add": {
            "post": {
                "description": "Create a new dish with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dishes (Admin)"
                ],
                "summary": "Add a new dish",
                "parameters": [
                    {
                        "description": "Dish details",
                        "name": "dish",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.AddDishPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Dish created successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "403": {
                        "description": "Access forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "500": {
                        "description": "Failed to add dish",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            }
        },
        "/api/dishes/admin/delete/{id}": {
            "delete": {
                "description": "Remove a dish from the system by its ID",
                "tags": [
                    "Dishes (Admin)"
                ],
                "summary": "Delete a dish by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Dish ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dish deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid dish ID",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "403": {
                        "description": "Access forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "500": {
                        "description": "Failed to delete dish",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            }
        },
        "/api/dishes/admin/update": {
            "put": {
                "description": "Update the details of an existing dish",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dishes (Admin)"
                ],
                "summary": "Update dish details",
                "parameters": [
                    {
                        "description": "Dish details",
                        "name": "dish",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ChangeDishPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dish updated successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "403": {
                        "description": "Access forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "500": {
                        "description": "Failed to update dish",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            }
        },
        "/api/dishes/by_category": {
            "post": {
                "description": "Retrieve a list of dishes based on their category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dishes"
                ],
                "summary": "Get dishes by category",
                "parameters": [
                    {
                        "description": "Category details",
                        "name": "category",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.GetDishesByCategoryPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of dishes",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Dish"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "404": {
                        "description": "No dishes found",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve dishes",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            }
        },
        "/api/dishes/by_id/{dish_id}": {
            "get": {
                "description": "Retrieve a specific dish by its ID",
                "tags": [
                    "Dishes"
                ],
                "summary": "Get a dish by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Dish ID",
                        "name": "dish_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dish details",
                        "schema": {
                            "$ref": "#/definitions/models.Dish"
                        }
                    },
                    "400": {
                        "description": "Invalid dish ID",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "404": {
                        "description": "Dish not found",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve dish",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            }
        },
        "/api/dishes/search": {
            "get": {
                "description": "Search for dishes by their name",
                "tags": [
                    "Dishes"
                ],
                "summary": "Search dishes by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dish name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of matching dishes",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Dish"
                            }
                        }
                    },
                    "400": {
                        "description": "Query parameter 'name' is required",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "404": {
                        "description": "No dishes found",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "500": {
                        "description": "Failed to search dishes",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.AddDishCategoryPayload": {
            "type": "object",
            "properties": {
                "category_name": {
                    "type": "string"
                }
            }
        },
        "handlers.AddDishPayload": {
            "type": "object",
            "required": [
                "dish_category",
                "dish_description",
                "dish_name",
                "dish_photo",
                "dish_price",
                "dish_weight"
            ],
            "properties": {
                "dish_category": {
                    "type": "integer"
                },
                "dish_description": {
                    "type": "string"
                },
                "dish_name": {
                    "type": "string"
                },
                "dish_photo": {
                    "type": "string"
                },
                "dish_price": {
                    "type": "number"
                },
                "dish_weight": {
                    "type": "number"
                }
            }
        },
        "handlers.ChangeDishPayload": {
            "type": "object",
            "required": [
                "dish_category",
                "dish_description",
                "dish_name",
                "dish_photo",
                "dish_price",
                "dish_weight",
                "id"
            ],
            "properties": {
                "dish_category": {
                    "type": "integer"
                },
                "dish_description": {
                    "type": "string"
                },
                "dish_name": {
                    "type": "string"
                },
                "dish_photo": {
                    "type": "string"
                },
                "dish_price": {
                    "type": "number"
                },
                "dish_weight": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "handlers.GetDishesByCategoryPayload": {
            "type": "object",
            "required": [
                "dish_category"
            ],
            "properties": {
                "dish_category": {
                    "type": "string"
                }
            }
        },
...
        "models.APIError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.Dish": {
            "type": "object",
            "required": [
                "dish_description",
                "dish_name",
                "dish_photo_url",
                "dish_price",
                "dish_weight"
            ],
            "properties": {
                "dish_category": {
                    "type": "string"
                },
                "dish_description": {
                    "type": "string"
                },
                "dish_name": {
                    "type": "string"
                },
                "dish_photo_url": {
                    "type": "string"
                },
                "dish_price": {
                    "type": "number"
                },
                "dish_rating": {
                    "type": "integer"
                },
                "dish_weight": {
                    "type": "number"
                }
            }
        },
...
    }
}